requirements.txt
pytorch
torchvision
scipy
kornia
tqdm